+++ projects/core/src/main/kotlin/site/siredvin/turtlematic/computercraft/peripheral/forged/ExperienceAutomataCorePeripheral.kt
@@ -2,6 +2,7 @@ package site.siredvin.turtlematic.computercraft.peripheral.forged
 
 import dan200.computercraft.api.turtle.ITurtleAccess
 import dan200.computercraft.api.turtle.TurtleSide
+import net.minecraft.world.entity.player.Player
 import site.siredvin.turtlematic.api.IAutomataCoreTier
 import site.siredvin.turtlematic.common.configuration.TurtlematicConfig
 import site.siredvin.turtlematic.computercraft.operations.SphereOperation
@@ -10,6 +11,7 @@ import site.siredvin.turtlematic.computercraft.peripheral.automatas.BaseAutomata
 import site.siredvin.turtlematic.computercraft.plugins.AutomataInteractionPlugin
 import site.siredvin.turtlematic.computercraft.plugins.AutomataItemSuckPlugin
 import site.siredvin.turtlematic.computercraft.plugins.AutomataLookPlugin
+import site.siredvin.tweakium.modules.peripheral.api.InteractionMode
 import site.siredvin.tweakium.modules.peripheral.boon.ExperienceBoon
 import site.siredvin.tweakium.modules.peripheral.boon.PeripheralOwnerBoonKey
 import site.siredvin.tweakium.modules.peripheral.boon.ScanningBoon
@@ -29,7 +31,9 @@ abstract class ExperienceAutomataCorePeripheral(
         peripheralOwner.attachBoon(PeripheralOwnerBoonKey.EXPERIENCE, ExperienceBoon(peripheralOwner, tier.interactionRadius, TurtlematicConfig.xpToFuelRate, UnconditionalOperation.XP_TRANSFER))
         addPlugin(AutomataItemSuckPlugin(this))
         addPlugin(AutomataLookPlugin(this))
-        addPlugin(AutomataInteractionPlugin(this))
+        addPlugin(AutomataInteractionPlugin(this, { it !is Player }, allowedMods = setOf(
+            InteractionMode.BLOCK, InteractionMode.ENTITY, InteractionMode.ANY
+        )))
         peripheralOwner.attachBoon(
             PeripheralOwnerBoonKey.SCANNING,
             ScanningBoon(