+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ability to scan, look and interact to forged automatas
+
 ## [1.4.5] - 2025-10-19
 
 ### Added